move inverse to own module
